Polymorphic Profile types!
Entities now have a list of profiles, and a Generic profile/info-type
will be added in the next few commits. This makes the /profile view
much easier to manage - it can just loop through the profiles and
output the json for each one (see `tentd/blueprints/entity.py`).

There's a lot of changes in this commit (sorry!), but once further
work is done it'll allow for generic profile types (see #26).

* Only import modules in tentd.models, not classes
* Added AppTestCase.commit, and removed tabs in that file